Refactored and categorised default settings
