+++ tardis/default_settings/filters.py
@@ -1,0 +1,13 @@
+# Post Save Filters
+# POST_SAVE_FILTERS = [
+#    ("tardis.tardis_portal.filters.exif.make_filter",
+#     ["EXIF", "http://exif.schema"]),  # this filter requires pyexiv2
+#                                       # http://tilloy.net/dev/pyexiv2/
+#    ]
+
+# Post Save Filters
+# POST_SAVE_FILTERS = [
+#    ("tardis.tardis_portal.filters.diffractionimage.make_filter",
+#     ["DIFFRACTION", "http://www.tardis.edu.au/schemas/trdDatafile/1",
+#      "/Users/steve/Desktop/diffdump"]),  #  requires ccp4 diffdump binary
+#    ]
