+++ tardis/default_settings/localisation.py
@@ -1,0 +1,20 @@
+USE_TZ = True
+
+# Local time zone for this installation. Choices can be found here:
+# http://en.wikipedia.org/wiki/List_of_tz_zones_by_name
+# although not all choices may be available on all operating systems.
+# If running in a Windows environment this must be set to the same as your
+# system time zone.
+
+TIME_ZONE = 'Australia/Melbourne'
+
+# Language code for this installation. All choices can be found here:
+# http://www.i18nguy.com/unicode/language-identifiers.html
+
+LANGUAGE_CODE = 'en-us'
+
+# Date format to use by default. ("jS F Y" => "8th March 2012")
+# https://docs.djangoproject.com/en/1.3/ref/templates/builtins/#std:templatefilter-date  # noqa
+
+DATE_FORMAT = "jS F Y"
+DATETIME_FORMAT = "jS F Y H:i"
