+++ tardis/default_settings/sharing.py
@@ -1,0 +1,3 @@
+TOKEN_EXPIRY_DAYS = 30
+TOKEN_LENGTH = 30
+TOKEN_USERNAME = 'tokenuser'
