+++ tardis/default_settings/site_customisations.py
@@ -1,0 +1,13 @@
+SITE_ID = 1
+
+SITE_TITLE = 'MyTardis'
+'''
+customise the title of your site
+'''
+
+SPONSORED_TEXT = None
+'''
+add text to the footer to acknowledge someone
+'''
+
+DEFAULT_INSTITUTION = "Monash University"
